pyproject.toml
[tool.poetry]
name = "repl_nix_gptbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
disnake = "^2.8.1"
init = "^0.1.0"
python-dotenv = "^1.0.0"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
